Fix GCC warnings

+++ test/custom_builds/c/timer/src/main.c
@@ -1,7 +1,7 @@
 #include <timer.h>
 
 typedef struct {
-    float x, y;
+    int32_t x, y;
 } Position, Velocity;
 
 void Move(ecs_iter_t *it) {
